Issue #8: Reimplementing GVariantType as a boxed type.
GVariantType is part of GLIB, but it is necessary to implement GVariantType
as a boxed type. Therefore we need parts of the GObject implementation.
Now GVariantType is implemented in a way to have the symbols in GLIB,
but it is loaded and compiled from GObject.

The tests for GIO now gives no errors. The tutorial example for an simple
application is running again.