Thread safety for stable-pointers (#36)
* Wrapping stable-pointer accesses with SBCL mutex

This works for me, in the sense that see no errors where nil gets funcall:ed. But it's a very naive approach, as I don't have a lot of experience with multi-threading.

* Wrapping stable-pointer accesses in mutex-locks.

This seems to work for me, in the sense that I don't see errors where nil gets funcall:ed. But I don't know if it might create other problems. Bordeaux-threads has to be loaded.

* Changed name sp-mutex -> stable-pointers lock

* bordeaux-threads now used by stable-pointers